LIVRABLES_PROJET10_Cassez_Guillaume_012026/1_Cassez_Guillaume_1_application_122024/azure_function/requirements.txt
# Azure Functions Core
azure-functions==1.18.0

# Azure Storage (équivalent de azure-storage-blob pour Azure Blob Storage)
azure-storage-blob==12.19.0

# Machine Learning & Data Science
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.3.0
pandas==2.0.3

# Utilities
python-dateutil==2.8.2
